[SOURCE]  * Driver SPI => OK  * Driver FM25H20 => OK
[TI_LIB]
 * Modification du linker script pour avoir plus de memoire

+++ include/FM25H20.h
@@ -43,8 +43,8 @@ extern "C"
 /* Public functions ------------------------------------------------------------------------------------------------*/
 void FM25H20_init(void);
 uint16_t FM25H250_statusRead(void);
-uint16_t FM25H20_memWrite(uint16_t address, const uint16_t *pData, uint32_t size);
-uint16_t FM25H20_memRead(uint16_t address, uint16_t *pData, uint32_t size);
+uint16_t FM25H20_memWrite(uint32_t address, const uint16_t *pData, uint32_t size);
+uint16_t FM25H20_memRead(uint32_t address, uint16_t *pData, uint32_t size);
 
 #endif /* INCLUDE_FM25H20_H_ */
